Fixes a race condition when Livewire uploads is used on a high-traffic website

+++ src/WithFileUploads.php
@@ -96,9 +96,11 @@ trait WithFileUploads
         $storage = FileUploadConfiguration::storage();
 
         foreach ($storage->allFiles(FileUploadConfiguration::path()) as $filePathname) {
-            $yesterdaysStamp = now()->subDay()->timestamp;
-            if ($yesterdaysStamp > $storage->lastModified($filePathname)) {
-                $storage->delete($filePathname);
+            if ($storage->exists($filePathname)) {
+                $yesterdaysStamp = now()->subDay()->timestamp;
+                if ($yesterdaysStamp > $storage->lastModified($filePathname)) {
+                    $storage->delete($filePathname);
+                }
             }
         }
     }
